+++ MastersThesisPOC/IAlgorithm.cs
@@ -2,6 +2,6 @@
 {
     public interface IAlgorithm
     {
-        (float result, string newMantissa, float delta) RunAlgorithm(string pattern, uint mantissa, int M, float x);
+        (float result, string newMantissa, float delta) RunAlgorithm(string pattern, uint mantissa, float M, float x);
     }
 }
